refactor(frontend): type-bound contracts, modularize feeds service, and refine result a11y
